Damage is now stored as an Integer, not just a short
-this means using the sparse OctreeObject, instead of OctreeShort, which should improve look-up times since there are rarely many of these with non-zero damage, and there is almost never any uniformity within the data which would benefit from the OctreeShort
-this also provides greater resolution for the possible damage values, being bound by 2^31 instead of 2^15 (as the maximum of 32 seconds was too low)
-note that this required adding a special-case to the decoding since we are now changing a serialized octree type in this version

+++ october-common/src/main/java/com/jeffdisher/october/data/IBlockProxy.java
@@ -29,7 +29,7 @@ public interface IBlockProxy
 	/**
 	 * @return The damage value for this block (usually 0).
 	 */
-	short getDamage();
+	int getDamage();
 	/**
 	 * @return The crafting operation for this block (usually null).
 	 */
